Support run with no search strategy (#132)
This PR adds support for running passes without search strategy. If
`search_stategy` is `None`, the passes are run on the input_model and
the final output model is evaluated if there is a valid evaluator. The
output of the run is the final output model and evaluation results.

## Tests 
Added tests for the new functionalities introduced. 
Updated relevant pre-existing tests.