Change `default_to_search` -> `disable_search` (#135)
This PR changes `default_to_search` to `disable_search` to make the
purpose of the parameter clear. When `disable_search` is `False`, the
pass config parameters uses the searchable values (if any) while they
use the default fixed values if `True`. The default value for
`disable_search` is `False`.

In `olive.workflows.run`, if there is a search strategy, the passes have
`disable_search` as `False` unless the user explicitly sets it to
`True`. If there search strategy is None, `disable_search` is always
`True` and user cannot override it.