Pareto frontier: check for `<=` instead of `<` (#165)
Currently, when checking if a point is dominated or not, we check if all
values for current points are less than those for another points.
However, this adds points to the pareto frontier that should not be
present. For instance, points `[1,1]`, `[2,1]` are both present even
though `[2,1]` is better. So we replace with `<=` and also check for
equality.

Mathematically speaking, `<` might be more correct but it looks like
`<=` is more used in practice. Optuna also does the same
https://github.com/optuna/optuna/blob/master/optuna/study/_multi_objective.py#L72

Before 

![less_than](https://user-images.githubusercontent.com/94929125/231330191-c18aa818-978f-44fd-a82f-164c46a70abe.png)

After

![less_than_or_equal](https://user-images.githubusercontent.com/94929125/231330229-5b6af153-bfd3-45cf-9e66-55dda03fc036.png)