Support external data for all onnx passes (#173)
Due the limit of protobuf, large onnx models (>2GB) cannot be saved to a
single onnx file. The tensors must be saved to an external data file.
Our onnx related passes do not handle this case.

This PR adds options for user to specify whether to save as external
data or not. If the model is too large, we override the user's
configuration and save as external data regardless. A warning is also
generated to inform the user of this. This ensures that the pass doesn't
fail when saving the onnx model to file.

The model save logic is based on a similar implementation by onnxruntime
https://github.com/microsoft/onnxruntime/blob/main/onnxruntime/python/tools/transformers/onnx_model.py#L947
but with more control on the name of the external data file.