Add io_config, dummy_inputs_func and dynamic axes to PytorchModel (#199)
Currently, model specific information such as `input_names`,
`input_shapes`, etc. are provided as config parameters for passes like
`ONNXConversion`. We want to support composite models which are made of
multiple models. For such cases, we cannot provide the io information to
a pass for all at once.

So, we make the io information optional init parameters of PytorchModel.
The passes will in turn get them from the model itself. This also makes
more sense conceptually since these are attributes of the model and not
the pass.

We will make similar changes to the ONNXModel and onnx related passes in
the future. Any model specific information should be part of the model
and not the pass config.