Differentiate SD fusions for DML EP based on version (#224)
DML in ORT 1.15.0 supports a subset of the transformer fusions, which
will be expanded in ORT 1.15.1 to include MHA. This change makes the
fusion options more explicit for DML EP based on ORT version. Some
fusion options are supported but disabled anyway since it limits the
number of contrib ops that don't provide a direct perf benefit for DML
EP.

Note that we still need DML 1.11.0 to be published and used in ORT
1.15.0 before nightly ORT-DML python packages can be used in Olive. I'll
convert this draft PR to a full PR once we have a nightly build to
consume.