Change `read_timeout` for `azure.ai.ml.MLClient` (#246)
## Describe your changes
The default `read_timeout` value for `MLClient` is `3000` seconds.
However, this value is too large and causes `AzureMLSystem` pass runs
and model evaluations to hang for a long time during the job streaming
and download steps when there are multiple jobs being created at the
same time (like in the CI pipelines). We already have retries set up for
this step, so it is better to reduce the read_timeout value for Olive.

With this change the bert and resnet ci tests now only take ~50 minutes
instead of ~2 hours it takes when the hanging happens.

We also allow user to set their own default values if they find the
connections are timing out even with retries. I don't expect this to be
needed much but it is safer to allow the user to overwrite instead of
hardcoding it to `60`.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Format your code by running `pre-commit run --all-files`

## (Optional) Issue link