Always try save_as_external_data for `model_proto_to_file`, handle external data for quant preprocess (#257)
## Describe your changes
When we try to save large onnx models (>2GB), we get a value error.
However, the error message depends on the version of `onnx`. We
previously tried to check the model byte size but there are cases when
the byte size returned is negative.
So, we always try again with `save_as_external_data=True` if the save
failed without it.
If this also fails, then raise original error. 

The `_quant_preprocess` method in onnx quantization pass doesn't handle
cases where the quant preprocess fails but the original model has
external data. In such cases, just copying the `.onnx` file is not
enough.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`

## (Optional) Issue link