Rethrow ImportError, Fix example tests output check, Update bert and whisper requirements (#350)
## Describe your changes
Add ImportError to list of exceptions to rethrow.  

Fix the output check for example tests. We need to check that
`footprints` is not empty. Otherwise, we don't catch failed engine runs
that just return empty footprints dict.

Updated the docker and azureml requirements so that the tests are not
broken anymore.
Updated the whisper example requirement to use onnxruntime-extensions
directly from pypi. Test is also updated to run `test_transcription`
script also.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link