Remove `use_distributed_tuning` config param from INC quantization passes (#375)
## Describe your changes
The newly release version of `neural-compressor` has removed the
`use_distributed_tuning` from their quant configs.
Since we were not using `use_distributed_tuning` in any of our examples
before, removing this doesn't affect any workflows.

Another option is to check for neural-compressor version and ignore this
parameter for the latest versions.

I think it is better to remove this parameter for now. In the latest
`neural-compressor`, distributed tuning must be done by calling the
python script with mpirun. This is not possible in Olive since we call
the tuner directly in our pass run code. There is no provision to run
olive with mpirun. We will need to think through how to support inc
distribued tuning if we want to support it in the future.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link