Whisper: Add multilingual support, Updated with latest ORT (#374)
## Describe your changes
This PR adds support for the latest ORT version and multilingual
support.

Also added transformer optimization pass in the `cpu,int8` config. It
was previously ommited due to a bug in ORT which was already resolved in
version `1.15.0`.

We currently don't have tests for the latest ORT and multilingual
support since there are no publically available ort-nightly builds. Will
look into this in the future.

Fixes: #293 
Fixes: #341 

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link