Make Olive compatible with ORT 1.16.0 (nightly) (#380)
## Describe your changes
ORT main branch recently removed some functions in
https://github.com/microsoft/onnxruntime/commit/89f8f20a61ceac0276c42d894c38400d86b175d0.

For vitis ai code, the import of these functions is made lazy (only
imported where needed). vitis ai code needs to be updated later to
actually be compatible with ORT 1.16.0. Since this version of ORT is not
released yet, it is not urgent.

For onnx quantization passes, `optimize_model` config parameter is
removed in ORT and we do the same. Using this option was already not
recommended so I think it might be okay to remove it from Olive onnx
quantization passes and always set it to `False` in older versions of
ORT. As described in the code comments, we already have
`quant_preprocess` which does the same optimization step as part of
preprocessing.
The commit above also says "This PR also remove the optimization from
quantization. User needs to call pre-process to optimize the model".

Whisper example requirements updated to `onnx==1.14.0` since the
workflow fails for ORT nightly otherwise. ORT versions 1.15.0,1.15.1 are
now compatible with onnx 1.14.0 because of
https://github.com/microsoft/Olive/commit/a531f77b87738c0916e08d123da74f30532a5a26.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link