Use pydantic<2.0.0 (#392)
## Describe your changes
There is a new major release for pydantic 2.0.0 which has multiple
breaking changes. So we pin pydantic to version < 2.0.0. We can consider
upgrading to the latest major release in the future.

There is no corresponding release for `autodoc_pydantic` yet but also
changing it to be safe.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link