Huggingface DataConfig: Add support for `data_files` and `max_samples` (#435)
## Describe your changes
This PR exposes the `data_files` argument for load_dataset. Sometimes we
only want specific files from the dataset to be used.

`max_samples` is used during pre-processing to limit the maximum number
of samples. Some datasets have a large number of samples which might be
too much for evaluation or calibration purposes.
In the future, we could provide other arguments such as
`max_calibration_samples` or `max_calibration_batches` if needed.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link