`text-generation` task support and `Perplexity` metric (#444)
## Describe your changes
This PR adds support for `text-generation` task in huggingface dataset.
Added new pre-processing and post-processing component for the task. The
pre-preprocessing component the tokenizes the input text and splits it
into chunks using randomization, non-overlapping blocks or sliding
window depending on the arguments provided.

Also adds Perplexity metric under accuracy type. 

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link