Keep `preds` and `targets` as torch tensor during accuracy evaluation (#451)
## Describe your changes
During accuracy evaluation, we convert the output and target tensors to
list before computing the metric value. However, this method is
expensive and not efficient when the tensors are large since we end up
converting them back to torch tensors anyways.
This is felt significantly when evaluating perplexity for large language
models where the `preds` tensor is a large logit tensor of size
`batch_size, seqlen, vocab` (like `1 X 2048 X 32000`). This is very
slow.

In this PR, we just move the tensors to `cpu` if not already on cpu and
concatenate them using `torch.cat`. This has the same effect as
`list.extend`.


## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link