Add `component_kwargs` in `DataConfig.params_config` (#454)
## Describe your changes
`DataConfig` supports multiple ways to pass parameters to the
components. They are:
- `DataConfig.params_config` - the parameters for different components
are provided as members of this dictionaries. Only the named parameters
of components are updated using this. There is no way to specify
component specific kwargs.
- `DataConfig.components["component_name"].params` - We can provide
component specific kwargs using this but it is hard to access. Moreover,
there is not way to provide this for hf datatemplate or
`HFConfig.datasets` for input model.

This PR introduces a reserved member in `DataConfig.params_config`
called `component_kwargs` where we can provide component specific
kwargs. An example usage is:
```
{
    "name": "c4_train",
    "type": "HuggingfaceContainer",
    "params_config": {
        "model_name": "gpt2",
        "task": "text-generation",
        "data_name": "allenai/c4",
        "subset": "allenai--c4",
        "split": "train",
        "data_files": {"train": "en/c4-train.00000-of-01024.json.gz"},
        "input_cols": ["text"],
        "component_kwargs": {
            "pre_process_data": {
                "max_samples": 128,
                "random_seed": 42,
                "seqlen": 2048,
            }
        }
    }
  }
```
This has automatic support for all templates and `HFConfig.dataset`. 

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link