Automatic onnx conversion for opt, llama, gptneox. Infer `feature` from `task` (#461)
## Describe your changes
Olive uses `transformer.onnx.FeaturesManager` to automatically get onnx
conversion args. However, huggingface has started using `optimum` for
this instead and haven't added support for newer models in
`transformers.onnx`.

In this PR, we add support for `opt`, `llama` and `gptneox` (model type
for dolly) by extending `FeaturesManager`. All of these models are text
decoder model so they are using the same onnx config.

Also, infer `feature` from `task` if not provided. We previously used
`default` but that doesn't work as expected if the `task` is different
from the default. Such as `sequence-classification` for gpt2 whose
default is `causal-lm`.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link