Separate `failed` and `invalid` pass run configs (#481)
## Describe your changes
Separate `PRUNED_CONFIG` into `INVALID_CONFIG` and `FAILED_CONFIG`. 
`INVALID_CONFIG` is used when the run config is invalid for a given
accelerator spec.
`FAILED_CONFIG` is used when the run config is valid, but the pass run
fails for some reason.

This helps us make the search more efficient. the transformer optimizer
pass was made `not accelerator_agnostic` even though the pass run itself
doesn't depend on the accelerator. All information is already in the
`run_config`.
It is okay for a pass to be `accelerator_agnostic` but the search space
to have points invalid for specific `accelerator_specs`.

What matters when determining if a pass is `accelerator_agnostic` is if
the pass run uses the accelerator specs and the output depends on it.
In the transformer optimizer case, the `use_gpu` information already
contains the necessary accelerator information so the run config has all
information needed.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link