Synchronize GPU and use cuda events for pytorch GPU latency (#492)
## Describe your changes
CUDA calls are asynchronous so using cpu timers like `time.perf_counter`
is not always correct (unless there are parts of the forward pass that
causes the GPU to synchronize before returning). It measures the time
taken to make the cuda kernal calls and not the entire execution time.

We use `torch.cuda.synchronize` after the forward call to wait for all
kernels to execute. Also use `torch.cuda.Event` to mark start and end
points for the forward pass.
Refer to
https://deci.ai/blog/measure-inference-time-deep-neural-networks/ for
more details.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link