Add `TorchTRTConversion` pass (#493)
## Describe your changes
This PR adds `TorchTRTConversion` pass which converts `nn.Linear`
modules in transformer layers to `TRTModule` from `torch-tensorrt`.
`torch-tensorrt` is an extension to `torch` where TensorRT engines can
be used like normal `nn.Module`.

We require this conversion to take advantage of `2:4` sparsity. 

~~I have not added unit test for the pass since the pass requires GPU
and we don't have GPU support for ci tests yet. I will add a unit test
in a follow up PR that uses aml system to run the pass and evaluation on
GPU.~~

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
Add `TorchTRTConversion` pass. 

## (Optional) Issue link