Bugfix: Insert `input_model` data_config and don't overwrite other hf data_configs (#502)
## Describe your changes
Previously, `RunConfigs.data_configs` had some default configs. We
remove the default configs and only insert a data config for the input
model if it has `hf_config.dataset`.

There was also a bug where all HF container data configs were being
overwritten by the input model's dataset. But we don't want to do this.
This has been fixed by only updating the `model_name` and `task` of the
user provided hf data_configs if not already present.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link