Remove `Pass._requires_data_config`, resolve using `endswith`, don't auto-insert input model dataset (#506)
## Describe your changes
This PR is an update to
[#506](https://github.com/microsoft/Olive/pull/505) with changes
discussed offline.

`Pass` has a class attribute `_requires_data_config` that we use to add
a config parameter with type `DataConfig` and name `data_config`. We
also use this in workflow run config validation to resolve string name
references or auto insert input models.

However, this functionality is limited. We can only add one `DataConfig`
parameter with name `data_config` and single description. Currently, all
of the passes are using the same parameter definition.

We remove `_requires_user_config` class attribute from passes.
DataConfig parameters must be defined along with other parameters in
`_default_config`. Parameter names ending in `default_config` are
reserved for such parameters. We use this to find and resolve DataConfig
parameters in run config validation.

Auto-insertion of the input model (if it's a pytorch model with
hf_config.dataset) data config in pass configs is removed. This behavior
only applies to a subset of pytorch input models and might be unexpected
to the user. If the user wants to use the input model's data config,
they can do so by assigning the value for the parameter as
`INPUT_MODEL_DATA_CONFIG`. After this, it behaves like any other string
name resolution.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link