`test-generation` dataset: Define strategy for `corpus` datasets, introduce `pair` dataset type (#509)
## Describe your changes
Follow up to [the
discussion](https://github.com/microsoft/Olive/pull/444#discussion_r1275799367)
in https://github.com/microsoft/Olive/pull/444.

There are many ways text-generation datasets are tokenized. We currently
support multiple methods where the text are joined together and
tokenized with or without a sliding window. The different methods were
selected based on the combination of arguments.

As I suggested in the linked discussion, this PR formalizes the
different methods of tokenization into `corpus_strategy`. The text is
joined or processed individually based on the strategy.

We also introduce a second type of dataset for `text-generation` tasks
called `pair`. Here, the dataset is expected to consist of two text
columns: input and output. Such as prompt and completion in chatbot type
scenario. Note that this is not a seq2seq scenerio since the model is
still generative. The input and output are concatenated into a single
sequence. The input might be ignored during evaluation (loss, metrics,
etc) based on the arguments provided.

Added some config classes to maintain and validate the different kwargs
accepted by the preprocessor.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link