Move reusable code from `sparsegpt_utils` to common and hf utils (#512)
## Describe your changes
`sparsegpt_utils` has some mappings and helper functions that are useful
in other parts of Olive too. This PR refactors it and puts the reusable
code into `olive.common.utils` and `olive.model.hf_utils`.

Also cleaned up the imports in `hf_utils` by moving `transformers`
imports to top of file. `transformers` is an olive requirement so we
don't need to lazy import it unless it is a version specific import.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link