Assert that `dataloader_func` or `data_config` is present if required (#516)
## Describe your changes
Some passes support both `dataloader_func` and `data_config`. We don't
set them to required but most of the time, one is required. Check for
this in the pass run so that the error is caught early and is clear to
the user.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link