Add `ConfigWithExtraArgs` (#534)
## Describe your changes
This PR adds a new config base class called `ConfigWithExtraArgs` where
the config values not defined in the class fields are collected in a
field called `extra_args`. This behaves similar to `**kwargs` in
functionality. It is useful for config classes where we expose some
parameters but still want to accept extra arguments. The child classes
can add their own validators on `extra_args` if needed.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link