ConfigWithExtraArgs: Handle cases where extra_args is provided as None (#535)
## Describe your changes
This is a follow up to https://github.com/microsoft/Olive/pull/534. The
`gather_extra_args` validator did not consider the case where the value
could be provided as None explicitly in the config (such as after
serialization). In which case `values.pop("extra_args", {})` returns
`None` which is not iterable.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link