check whether the inference session is fallback after it is created (#540)
## Describe your changes
From the time being, the ORT will automatically allow execution provider
fallback. This behavior is not expected for Olive which targeted on the
specific EP optimization.

We explicitly check whether it is fallbacked or not. If yes, throw
exception, otherwise, we will disable the fallback.


## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link