Pipelines: Remove run-olive-script (#562)
## Describe your changes
Removes `run-olive-script` template since it's not useful anymore. We
don't use venv in the pipelines so no need to activate it.
`PIPELINE` argument for makefile is also removed.

Forced installation of cpu pytorch is also removed since torch is
already a dependency of olive. gpu pytorch should support all cpu
functionalities.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link