Move model preparation logic out of model itself (#561)
## Describe your changes
From the time being, the OliveModel is used in Olive to prepare the
optimization. It should use model config instead of model itself. For
example, if the model is hosted in AML, Engine need create the model
instance for run. later, it will need manually set the download model
path to model instance. This kind of behavior hinder the lifecycle of
the model. In case of remote model path scenario, we cannot call any
model path related method in constructor. Such as
https://github.com/microsoft/Olive/pull/542#discussion_r1320161984.

In this PR, I split the responsibility of model itself and model
preparation to different place. The assumption is whenever the model
instance is created, the model path should already be downloaded and can
directly access.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link