Don't warn for repeated data component registration (#601)
## Describe your changes
Data component registration warns if a component has already been
registered. However, this warning gets triggered every time a data
config is initialized. This happens many times since data configs get
converted to dict/json and back to object many times.
There is nothing the user can do to prevent this so it should not be a
warning. Changed to debug.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link