Remove duplicate values from pytorch model json (#600)
## Describe your changes
Pytorch model adds values from the hf model config into model attributes
to make them easily accessible. However, this makes the model json
messy/confusing with values in model_attributes that are already in
`model_config`. This PR removes information from model_attributes that
is already in the hf_config.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link