ResNet: Update seed_everything and accuracy goal (#646)
## Describe your changes
Updated the seed steps to make the model consistent accross runs. 

Update the accuracy goal in `resnet_ptq_cpu_aml_dataset.json` to
increase the tolerance. This was not increased last time along with
`r`resnet_ptq_cpu_aml_dataset.json`` and was causing the test to be
flaky. Reduced the value for `resnet_ptq_cpu_aml_dataset.json` from
`0.1` to `0.05` since the baseline is better now.


## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link