Option to use custom root directory for tempfile (#683)
## Describe your changes
Olive uses a lot of temporary directories which are created in the
default tempory directory root of the system. However, we encounter
issues with storage, permissions, files disappearing, etc.
This PR adds a command line option `--tempdir` that changes the default
`dir` for tempfile for the run
https://docs.python.org/3/library/tempfile.html#tempfile.tempdir

For python script use cases, the user must set the value of
`tempfile.tempdir` themselves. Its's better for us to not have to deal
with this value dynamically.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link