[LlaMA2] Clean up and update scripts (#692)
## Describe your changes
### `llama2.py`
- Update description for `--use_gqa` and make it only valid for `gpu`.
- Add `--tempdir` option to specify custom tempdir root. Useful when the
default tempdir root is out of space.
- Version check for `onnxruntime>=1.16.2`
- Remove unused components from the generated configs to avoid confusion
- The above also corrected an issue where the cpu config still used
`gqa_evaluator` for some passes.
- Generated user_scripts are suffixed with `model_name` so that input
model cache can be reused between runs.

### `user_script.py`
- Removed unused functions for legacy `decoder` and `decoder_with_past`
models.
- Made a general `get_merged_sample_with_past_kv_inputs` function that
can be used to create dummy inputs for both legacy models and the new
"merged" model. The comments describe all possible use cases if needed
in the future.
- All inputs are generated as tensors on cpu. The evalutors handle the
device and cast to numpy as needed.
- GQA model uses shared kv buffer so the past key values need to be of
length `max_length`. **NOTE**: We currently do not use the correct
io_binding and buffer sharing since our evaluator io_binding support
doesn't support it. This needs to be updated later.
https://github.com/microsoft/onnxruntime/blob/d8d79521ca2b266e631ac0ba7036a682ebb58b5b/onnxruntime/python/tools/transformers/models/llama/llama_inputs.py#L292
- `RandomDataLoader` is updated to generate the inputs itself to keep it
simple.
- dataloaders don't ignore `batch_size` 

The following are the latency numbers from one A100 GPU:
**llama2_gpu_mha**

![llam2_gpu_mha](https://github.com/microsoft/Olive/assets/94929125/93a1240c-7b86-4117-a716-000be09ad9c8)

**llama2_gpu_gqa**

![llama2_gpu_gqa](https://github.com/microsoft/Olive/assets/94929125/9188afeb-ba13-4e4d-969f-bbe939ecafbc)
The gqa model's latency is not as good as expected but that is probably
due to the incorrect io_binding mentioned above.
The gqa op is also still under development and debugging so this can be
fixed later.


## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link